Fixed the bug where the watched selection() is not called at certain (#1945)
situations when the caret position changes. The actual caret
position will be moved to the end of input field particularly when the
mask changes dynamically, and it should be repositioned to the value
of this.selection. However, because this.selection is watched as
unchanged, selection() is not called to fix the caret position.