revert(v-tooltip): reverted #2206 change for tooltips
when a `v-tooltip` was used inside of a `v-radio-group`, it was not properly extending causing a
text overflow. While the fix worked, `v-radio-group` does not support direct children not being
`v-radio` components. reverting for now and pushing #2171 to v1.1.x

breaks change made in #2206 for #2171

(cherry picked from commit 426ce14bba7f77ca93f06a74c729625823540621)