fix(VDataTable): reset expanded value to object instead of array (#2666)
Clicking a VDataTable header triggers a reset of the expanded property.  The functionality requires an object, but this mixin sets it to an array.

This causes VDataTable item-key values that are numbers to create an array the size of the item-key value. What we want is an object with an item-key to bool mapping.