v-select with combobox option uses search value when blurring (#2879)
* Rename findExistingItem to findExistingIndex

Said function doesn't return an item, but rather the index of that item
within the items array. The name findExistingIndex reflects better what
the function actually does.

* fix: v-select with combobox option uses searchValue when blurring

When using v-select with the combobox option and items as objects (not
just primitives), and you select an option and then blur the field,
the itemValue becomes the text value of the selected item object.

fixes #2878

* Split comment into two lines

* Store the result of this.getValue(val) in a variable

This prevents it from beiing re-evaluated for every find iteration.