docs(VDivider): add a11y section (#7709)
* docs: add a11y section

Adds an optional section to UI doc pages for describing/explaining accessibility (a11y)
considerations. Added at the bottom of the Page template right before

#7694

* docs(VDivider): add a11y considerations

Example of the type of content that we might include in an Accessibility Considerations section for
UI docs pages.

#7694 #7695

* Revert "docs: add a11y section"

This reverts commit b3ce305353eb141bdb2607ffe8d02ed12df27a2e.

* code => text

Edited for consistency. Made any text that is not either a component or property name bold vs code.