docs: force consistent spelling of vue cli 3 (#8112)
Vue CLI 3 is spelling several ways across the docs, in part because Vue changed the package name of
the CLI between v2 and v3. This commit simply makes the spelling consistent throughout. Eventually,
it may be better to drop the version number entirely and just use 'Vue CLI' as the Vue team does.