docs(display.md + flex.md): add caveats section (#16080)
## Description
<!-- Describe your changes in detail -->
I have added information about the usage of !important in the display
classes. Such as how this could be an issue and what would happen with
an example noted.

<!-- Note any issues that are resolved by this PR -->
<!-- e.g. resolves #4213 or fixes #2312 -->
#14773

## Motivation and Context
<!-- Why is this change required? What problem does it solve? -->
This is a needed addition to the documentation because some people might
overlook that `!important` is added in the display classes.

<!-- If it fixes an open issue, please link to the issue here. -->
#14773

## How Has This Been Tested?
Since this is just documentation, it has been tested locally on my
computer, both with a dev server and after building.

## Markup:
N/A

## Types of changes
<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Improvement/refactoring (non-breaking change that doesn't add any
features but makes things better)

## Checklist:
<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!-- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->
- [x] The PR title is no longer than 64 characters.
- [x] The PR is submitted to the correct branch (`master` for bug fixes
and documentation updates, `dev` for new features and backwards
compatible changes and `next` for non-backwards compatible changes).
- [x] My code follows the code style of this project.
- [x] I've added relevant changes to the documentation (applies to new
features and breaking changes in core library)