fix(VDivider): fix color attribute (#16536)
Replaced useTextColor instead of useBackgroundColor.

resolves #16517

<!--
MAKE SURE TO READ THE CONTRIBUTING GUIDE BEFORE CREATING A PR
https://vuetifyjs.com/getting-started/contributing

Provide a general summary of your changes in the title above
Keep the title short and descriptive, as it will be used as a commit
message
PR titles should follow conventional-changelog-angular:
https://vuetifyjs.com/getting-started/contributing/#commit-guidelines
-->

## Description
To change divider color, we need to set the color property instead of
the background-color property. That's why I replaced useBackgroundColor
composable with useTextColor.
<!--
Describe your changes in detail. Why is this change required? What
problem does it solve?
If it fixes an open issue, please link to the issue here.
e.g. resolves #4213 or fixes #2312
-->

## Markup:

<!--
Information on how to set up your local development environment can be
found here:

https://vuetifyjs.com/getting-started/contributing/#setting-up-your-environment
Remove this section for documentation or test-only changes.
-->

<!-- Paste your FULL packages/vuetify/dev/Playground.vue here --->
```vue
<template>
  <v-container>
    <v-card color="" dark class="mx-auto" width="400px" hight="100vh" :loading="isUpdating">
      <v-divider color="#fff" thickness="1"></v-divider>
          <span>Some text</span>

          <v-divider color="rgb(0,0,0)" thickness="2"></v-divider>
          <span>Some text</span>
          <v-divider color="rgb(0,0,0)" thickness="3"></v-divider>
          <span>Some text</span>
          <v-divider color="#92fd93" thickness="4"></v-divider>
          <span>Some text</span>
          <v-divider color="rgb(0,0,0)" thickness="5"></v-divider>
          <span>Some text</span>
          <v-divider class="rgb(0,0,0)" thickness="6"></v-divider>
              <span>Some text</span>      
          <v-divider  class="text-pink-lighten-1" thickness="7"></v-divider>
          <span>Some text</span>
          <v-divider color="purple-darken-4" thickness="8"></v-divider>
    </v-card>
  </v-container>
</template>

```
#Types of changes
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Improvement/refactoring (non-breaking change that doesn't add any
features but makes things better)

#Checklist:
- [x] The PR title is no longer than 64 characters.
- [x] The PR is submitted to the correct branch (master for bug fixes
and documentation updates, dev for new features and backwards compatible
changes and next for non-backwards compatible changes).
- [x] My code follows the code style of this project.
- [x] I've added relevant changes to the documentation (applies to new
features and breaking changes in core library)